ABC base class for HandshakerAPI
