+++ newsfragments/1052.misc.rst
@@ -1,0 +1,1 @@
+Add ``p2p.abc.HandshakerAPI`` for formal handshaker interface.
